+++ src/main/java/net/neoforged/neoforge/common/data/internal/NeoForgeDataMapsProvider.java
@@ -10,7 +10,8 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.world.level.block.ComposterBlock;
 import net.neoforged.neoforge.common.data.DataMapProvider;
-import net.neoforged.neoforge.registries.datamaps.NeoForgeDataMaps;
+import net.neoforged.neoforge.registries.datamaps.builtin.Compostable;
+import net.neoforged.neoforge.registries.datamaps.builtin.NeoForgeDataMaps;
 
 public class NeoForgeDataMapsProvider extends DataMapProvider {
     public NeoForgeDataMapsProvider(PackOutput packOutput, CompletableFuture<HolderLookup.Provider> lookupProvider) {
@@ -20,6 +21,6 @@ public class NeoForgeDataMapsProvider extends DataMapProvider {
     @Override
     protected void gather() {
         final var compostables = builder(NeoForgeDataMaps.COMPOSTABLES);
-        ComposterBlock.COMPOSTABLES.forEach((item, chance) -> compostables.add(item.asItem().builtInRegistryHolder(), new NeoForgeDataMaps.Compostable(chance), false));
+        ComposterBlock.COMPOSTABLES.forEach((item, chance) -> compostables.add(item.asItem().builtInRegistryHolder(), new Compostable(chance), false));
     }
 }
